fix(byok): flat-key i18n format, always-on Base URL, move into Settings dialog
Three fixes for the BYOK (Bring Your Own Key) console page so the actual UI is usable:

1. **i18n showed raw keys (`byok.title.list`, `byok.field.provider`, ...)** — every `src/i18n/<lang>/byok.json` was authored in nested shape (`{title: {list: '...'}}` or `{title: {list: {message, description}}}`), but the loader at `src/i18n/index.ts` iterates flat top-level keys and reads `element.message`:

       messages[`${name}.${key}`] = element.message

   That returned `undefined` for every BYOK key. Flatten all 18 byok.json
   files to the correct shape (`{"section.key": {message, description}}`)
   while preserving the existing translated text — same shape every other
   namespace (e.g. `console.json`) already uses.

2. **Base URL is no longer hidden behind an 'advanced' toggle.** The
   `baseUrlHelp` text already explains it's optional, and the toggle just
   added a click of friction for the non-trivial number of users who use
   OpenRouter / a self-hosted reverse proxy / an enterprise gateway. Show
   the field directly under `API Key` like every other form field.

3. **Move the page out of `/console/byok` into the user Settings dialog
   (avatar dropdown → Settings → 'API Key' tab).** BYOK is per-user
   configuration, not an Application/Order-style console resource — it
   belongs alongside Theme/Language/Site/SEO/etc., not next to 'Order
   list' and 'Usage history'. Net result for the user: one obvious place
   to manage their personal API keys, no second sidebar to navigate.

   Implementation details:
   - New `src/components/setting/Byok.vue` is a slim wrapper around the
     existing list logic, sized to fit the settings dialog body.
   - Existing dialog `src/components/console/byok/Dialog.vue` moves to
     `src/components/setting/byok/Dialog.vue` (kept the implementation
     untouched apart from the advanced-toggle removal above).
   - `UserSetting.vue` widens the dialog to 900px when the BYOK tab is
     active so the credential table doesn't overflow.
   - Drops the `/console/byok` route, `ROUTE_CONSOLE_BYOK_LIST` constant,
     console sidebar link, and mobile-profile link. Old direct links
     to `/console/byok` will 404 — acceptable since the entry was added
     in #555 and isn't in production for long.
   - Token bootstrap (lazy-load chat module → application →
     `chat.credential.token`) is preserved in the new component so the
     settings dialog still authenticates against `/aichat2/credentials`.

The orphan `console.menu.byokList` translation key is left in place
across all locales; harmless and removing it from 18 files would balloon
this PR's noise.

Validation: `npx vue-tsc --noEmit` clean, `npm run lint` clean. Manual
smoke needed against a deployed build (this can't easily be exercised in
CI because it needs a chat-Application credential).
