+++ src/i18n/en/common.json
@@ -703,6 +703,10 @@
     "message": "Function Settings",
     "description": "The title of the function settings section in the settings page."
   },
+  "settings.apiKey": {
+    "message": "API Key",
+    "description": "Sidebar label in the settings dialog for managing user-supplied (BYOK) upstream provider API Keys."
+  },
   "nav.gemini": {
     "message": "Gemini",
     "description": "Text in the website navigation bar to display the Gemini page, must remain as 'Gemini'."
