+++ src/i18n/es/byok.json
@@ -1,54 +1,154 @@
 {
-  "title": {
-    "list": "Clave API personalizada",
-    "addCredential": "Agregar clave API",
-    "editCredential": "Editar clave API",
-    "deleteCredential": "Eliminar clave API"
+  "title.list": {
+    "message": "Clave API personalizada",
+    "description": ""
   },
-  "description": {
-    "list": "Utiliza tu propia clave API de proveedor (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) en lugar del pool de cuentas de la plataforma. Una vez configurado, las solicitudes de conversación del modelo correspondiente se realizarán directamente con tu clave a los proveedores, sin consumir créditos de la plataforma; las funciones de herramientas, referencias, artefactos, historial en la nube y otras seguirán funcionando normalmente."
+  "title.addCredential": {
+    "message": "Agregar clave API",
+    "description": ""
   },
-  "field": {
-    "provider": "Proveedor",
-    "label": "Notas",
-    "labelPlaceholder": "Ej. \"Mi número de trabajo\" (opcional)",
-    "apiKey": "Clave API",
-    "apiKeyPlaceholder": "sk-...",
-    "apiKeyEditPlaceholder": "Dejar vacío para mantener la clave actual",
-    "baseUrl": "URL base personalizada",
-    "baseUrlPlaceholder": "Dejar vacío para usar la dirección predeterminada del proveedor",
-    "isActive": "Habilitado",
-    "lastUsedAt": "Último uso",
-    "createdAt": "Fecha de adición"
+  "title.editCredential": {
+    "message": "Editar clave API",
+    "description": ""
   },
-  "value": {
-    "active": "Habilitado",
-    "inactive": "No habilitado",
-    "never": "Nunca usado",
-    "default": "Predeterminado"
+  "title.deleteCredential": {
+    "message": "Eliminar clave API",
+    "description": ""
   },
-  "button": {
-    "add": "Agregar clave API",
-    "edit": "Editar",
-    "test": "Probar",
-    "delete": "Eliminar",
-    "save": "Guardar",
-    "cancel": "Cancelar"
+  "description.list": {
+    "message": "Utiliza tu propia clave API de proveedor (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) en lugar del pool de cuentas de la plataforma. Una vez configurado, las solicitudes de conversación del modelo correspondiente se realizarán directamente con tu clave a los proveedores, sin consumir créditos de la plataforma; las funciones de herramientas, referencias, artefactos, historial en la nube y otras seguirán funcionando normalmente.",
+    "description": ""
   },
-  "message": {
-    "deleteConfirm": "¿Confirmar la eliminación de esta clave API? Las conversaciones del modelo {provider} volverán a la configuración predeterminada de la plataforma.",
-    "deleteSuccess": "Clave API eliminada.",
-    "saveSuccess": "Clave API guardada.",
-    "testSuccess": "Conexión exitosa.",
-    "testFailed": "Prueba fallida: {message}",
-    "loadFailed": "Error al cargar las credenciales.",
-    "saveFailed": "Error al guardar las credenciales.",
-    "empty": "Aún no se ha configurado una clave API personalizada. Haz clic en \"Agregar clave API\" para usar tu propia cuenta de proveedor.",
-    "advancedToggle": "Mostrar opciones avanzadas",
-    "baseUrlHelp": "Opcional. Dirección personalizada de upstream (para OpenRouter, proxy inverso propio, gateway empresarial, etc.). Solo se permite HTTPS, no puede apuntar a direcciones de red interna/loopback."
+  "field.provider": {
+    "message": "Proveedor",
+    "description": ""
   },
-  "badge": {
-    "active": "Usando tu clave {provider}",
-    "tooltip": "Esta conversación utiliza tu propia clave API de {provider}, los costos serán cobrados a la cuenta de {provider}, sin descontar el saldo de la plataforma."
+  "field.label": {
+    "message": "Notas",
+    "description": ""
+  },
+  "field.labelPlaceholder": {
+    "message": "Ej. \"Mi número de trabajo\" (opcional)",
+    "description": ""
+  },
+  "field.apiKey": {
+    "message": "Clave API",
+    "description": ""
+  },
+  "field.apiKeyPlaceholder": {
+    "message": "sk-...",
+    "description": ""
+  },
+  "field.apiKeyEditPlaceholder": {
+    "message": "Dejar vacío para mantener la clave actual",
+    "description": ""
+  },
+  "field.baseUrl": {
+    "message": "URL base personalizada",
+    "description": ""
+  },
+  "field.baseUrlPlaceholder": {
+    "message": "Dejar vacío para usar la dirección predeterminada del proveedor",
+    "description": ""
+  },
+  "field.isActive": {
+    "message": "Habilitado",
+    "description": ""
+  },
+  "field.lastUsedAt": {
+    "message": "Último uso",
+    "description": ""
+  },
+  "field.createdAt": {
+    "message": "Fecha de adición",
+    "description": ""
+  },
+  "value.active": {
+    "message": "Habilitado",
+    "description": ""
+  },
+  "value.inactive": {
+    "message": "No habilitado",
+    "description": ""
+  },
+  "value.never": {
+    "message": "Nunca usado",
+    "description": ""
+  },
+  "value.default": {
+    "message": "Predeterminado",
+    "description": ""
+  },
+  "button.add": {
+    "message": "Agregar clave API",
+    "description": ""
+  },
+  "button.edit": {
+    "message": "Editar",
+    "description": ""
+  },
+  "button.test": {
+    "message": "Probar",
+    "description": ""
+  },
+  "button.delete": {
+    "message": "Eliminar",
+    "description": ""
+  },
+  "button.save": {
+    "message": "Guardar",
+    "description": ""
+  },
+  "button.cancel": {
+    "message": "Cancelar",
+    "description": ""
+  },
+  "message.deleteConfirm": {
+    "message": "¿Confirmar la eliminación de esta clave API? Las conversaciones del modelo {provider} volverán a la configuración predeterminada de la plataforma.",
+    "description": ""
+  },
+  "message.deleteSuccess": {
+    "message": "Clave API eliminada.",
+    "description": ""
+  },
+  "message.saveSuccess": {
+    "message": "Clave API guardada.",
+    "description": ""
+  },
+  "message.testSuccess": {
+    "message": "Conexión exitosa.",
+    "description": ""
+  },
+  "message.testFailed": {
+    "message": "Prueba fallida: {message}",
+    "description": ""
+  },
+  "message.loadFailed": {
+    "message": "Error al cargar las credenciales.",
+    "description": ""
+  },
+  "message.saveFailed": {
+    "message": "Error al guardar las credenciales.",
+    "description": ""
+  },
+  "message.empty": {
+    "message": "Aún no se ha configurado una clave API personalizada. Haz clic en \"Agregar clave API\" para usar tu propia cuenta de proveedor.",
+    "description": ""
+  },
+  "message.advancedToggle": {
+    "message": "Mostrar opciones avanzadas",
+    "description": ""
+  },
+  "message.baseUrlHelp": {
+    "message": "Opcional. Dirección personalizada de upstream (para OpenRouter, proxy inverso propio, gateway empresarial, etc.). Solo se permite HTTPS, no puede apuntar a direcciones de red interna/loopback.",
+    "description": ""
+  },
+  "badge.active": {
+    "message": "Usando tu clave {provider}",
+    "description": ""
+  },
+  "badge.tooltip": {
+    "message": "Esta conversación utiliza tu propia clave API de {provider}, los costos serán cobrados a la cuenta de {provider}, sin descontar el saldo de la plataforma.",
+    "description": ""
   }
 }
