+++ src/i18n/fi/byok.json
@@ -1,54 +1,154 @@
 {
-  "title": {
-    "list": "Mukautettu API-avain",
-    "addCredential": "Lisää API-avain",
-    "editCredential": "Muokkaa API-avain",
-    "deleteCredential": "Poista API-avain"
+  "title.list": {
+    "message": "Mukautettu API-avain",
+    "description": ""
   },
-  "description": {
-    "list": "Käytä omaa palveluntarjoajan API-avain (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) sen sijaan, että käyttäisit alustan tilipoolia. Konfiguroinnin jälkeen vastaavat mallin keskustelupyynnöt kutsuvat suoraan ylävirran palveluntarjoajaa avaimellasi, eikä se kuluta alustan krediittejä; työkalukutsut, viittaukset, artefaktit, pilvihistoria ja muut toiminnot toimivat normaalisti."
+  "title.addCredential": {
+    "message": "Lisää API-avain",
+    "description": ""
   },
-  "field": {
-    "provider": "Palveluntarjoaja",
-    "label": "Huomautus",
-    "labelPlaceholder": "esim. \"Työnumero\" (valinnainen)",
-    "apiKey": "API-avain",
-    "apiKeyPlaceholder": "sk-...",
-    "apiKeyEditPlaceholder": "Jätä tyhjäksi, jotta nykyinen avain pysyy muuttumattomana",
-    "baseUrl": "Mukautettu perus-URL",
-    "baseUrlPlaceholder": "Jätä tyhjäksi käyttääksesi palveluntarjoajan oletusosoitetta",
-    "isActive": "Aktiivinen",
-    "lastUsedAt": "Viimeksi käytetty",
-    "createdAt": "Lisäysaika"
+  "title.editCredential": {
+    "message": "Muokkaa API-avain",
+    "description": ""
   },
-  "value": {
-    "active": "Aktiivinen",
-    "inactive": "Ei aktiivinen",
-    "never": "Ei koskaan käytetty",
-    "default": "Oletus"
+  "title.deleteCredential": {
+    "message": "Poista API-avain",
+    "description": ""
   },
-  "button": {
-    "add": "Lisää API-avain",
-    "edit": "Muokkaa",
-    "test": "Testaa",
-    "delete": "Poista",
-    "save": "Tallenna",
-    "cancel": "Peruuta"
+  "description.list": {
+    "message": "Käytä omaa palveluntarjoajan API-avain (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) sen sijaan, että käyttäisit alustan tilipoolia. Konfiguroinnin jälkeen vastaavat mallin keskustelupyynnöt kutsuvat suoraan ylävirran palveluntarjoajaa avaimellasi, eikä se kuluta alustan krediittejä; työkalukutsut, viittaukset, artefaktit, pilvihistoria ja muut toiminnot toimivat normaalisti.",
+    "description": ""
   },
-  "message": {
-    "deleteConfirm": "Vahvistatko tämän API-avaimen poistamisen? Jatkossa {provider} mallin keskustelut palaavat alustan oletusylävirtaan.",
-    "deleteSuccess": "API-avain on poistettu.",
-    "saveSuccess": "API-avain on tallennettu.",
-    "testSuccess": "Yhteys on kunnossa.",
-    "testFailed": "Testi epäonnistui: {message}",
-    "loadFailed": "Todistusten lataaminen epäonnistui.",
-    "saveFailed": "Todistusten tallentaminen epäonnistui.",
-    "empty": "Mukautettua API-avainta ei ole vielä konfiguroitu. Napsauta \"Lisää API-avain\" käyttääksesi omaa palveluntarjoajan tiliäsi.",
-    "advancedToggle": "Näytä lisäasetukset",
-    "baseUrlHelp": "Valinnainen. Mukautettu ylävirran osoite (käytettävä OpenRouterissa, itse rakennetuissa käänteisissä proxyissä, yritysporttien jne. kanssa). Vain HTTPS sallittu, ei saa osoittaa sisäverkkoon/loopback-osoitteeseen."
+  "field.provider": {
+    "message": "Palveluntarjoaja",
+    "description": ""
   },
-  "badge": {
-    "active": "Käyttää omaa {provider} avaintasi",
-    "tooltip": "Tässä keskustelussa käytetään omaa {provider} API-avaintasi, kustannukset veloitetaan {provider} tililtä, eikä se vähennä alustan saldoa."
+  "field.label": {
+    "message": "Huomautus",
+    "description": ""
+  },
+  "field.labelPlaceholder": {
+    "message": "esim. \"Työnumero\" (valinnainen)",
+    "description": ""
+  },
+  "field.apiKey": {
+    "message": "API-avain",
+    "description": ""
+  },
+  "field.apiKeyPlaceholder": {
+    "message": "sk-...",
+    "description": ""
+  },
+  "field.apiKeyEditPlaceholder": {
+    "message": "Jätä tyhjäksi, jotta nykyinen avain pysyy muuttumattomana",
+    "description": ""
+  },
+  "field.baseUrl": {
+    "message": "Mukautettu perus-URL",
+    "description": ""
+  },
+  "field.baseUrlPlaceholder": {
+    "message": "Jätä tyhjäksi käyttääksesi palveluntarjoajan oletusosoitetta",
+    "description": ""
+  },
+  "field.isActive": {
+    "message": "Aktiivinen",
+    "description": ""
+  },
+  "field.lastUsedAt": {
+    "message": "Viimeksi käytetty",
+    "description": ""
+  },
+  "field.createdAt": {
+    "message": "Lisäysaika",
+    "description": ""
+  },
+  "value.active": {
+    "message": "Aktiivinen",
+    "description": ""
+  },
+  "value.inactive": {
+    "message": "Ei aktiivinen",
+    "description": ""
+  },
+  "value.never": {
+    "message": "Ei koskaan käytetty",
+    "description": ""
+  },
+  "value.default": {
+    "message": "Oletus",
+    "description": ""
+  },
+  "button.add": {
+    "message": "Lisää API-avain",
+    "description": ""
+  },
+  "button.edit": {
+    "message": "Muokkaa",
+    "description": ""
+  },
+  "button.test": {
+    "message": "Testaa",
+    "description": ""
+  },
+  "button.delete": {
+    "message": "Poista",
+    "description": ""
+  },
+  "button.save": {
+    "message": "Tallenna",
+    "description": ""
+  },
+  "button.cancel": {
+    "message": "Peruuta",
+    "description": ""
+  },
+  "message.deleteConfirm": {
+    "message": "Vahvistatko tämän API-avaimen poistamisen? Jatkossa {provider} mallin keskustelut palaavat alustan oletusylävirtaan.",
+    "description": ""
+  },
+  "message.deleteSuccess": {
+    "message": "API-avain on poistettu.",
+    "description": ""
+  },
+  "message.saveSuccess": {
+    "message": "API-avain on tallennettu.",
+    "description": ""
+  },
+  "message.testSuccess": {
+    "message": "Yhteys on kunnossa.",
+    "description": ""
+  },
+  "message.testFailed": {
+    "message": "Testi epäonnistui: {message}",
+    "description": ""
+  },
+  "message.loadFailed": {
+    "message": "Todistusten lataaminen epäonnistui.",
+    "description": ""
+  },
+  "message.saveFailed": {
+    "message": "Todistusten tallentaminen epäonnistui.",
+    "description": ""
+  },
+  "message.empty": {
+    "message": "Mukautettua API-avainta ei ole vielä konfiguroitu. Napsauta \"Lisää API-avain\" käyttääksesi omaa palveluntarjoajan tiliäsi.",
+    "description": ""
+  },
+  "message.advancedToggle": {
+    "message": "Näytä lisäasetukset",
+    "description": ""
+  },
+  "message.baseUrlHelp": {
+    "message": "Valinnainen. Mukautettu ylävirran osoite (käytettävä OpenRouterissa, itse rakennetuissa käänteisissä proxyissä, yritysporttien jne. kanssa). Vain HTTPS sallittu, ei saa osoittaa sisäverkkoon/loopback-osoitteeseen.",
+    "description": ""
+  },
+  "badge.active": {
+    "message": "Käyttää omaa {provider} avaintasi",
+    "description": ""
+  },
+  "badge.tooltip": {
+    "message": "Tässä keskustelussa käytetään omaa {provider} API-avaintasi, kustannukset veloitetaan {provider} tililtä, eikä se vähennä alustan saldoa.",
+    "description": ""
   }
 }
