+++ src/i18n/it/byok.json
@@ -1,54 +1,154 @@
 {
-  "title": {
-    "list": "Chiavi API personalizzate",
-    "addCredential": "Aggiungi chiave API",
-    "editCredential": "Modifica chiave API",
-    "deleteCredential": "Elimina chiave API"
+  "title.list": {
+    "message": "Chiavi API personalizzate",
+    "description": ""
   },
-  "description": {
-    "list": "Utilizza la tua chiave API del fornitore di servizi (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) al posto del pool di account della piattaforma. Una volta configurato, le richieste di dialogo per il modello corrispondente verranno effettuate direttamente con la tua chiave, senza consumare crediti della piattaforma; le chiamate agli strumenti, le citazioni, i lavori e altre funzionalità come la cronologia cloud funzioneranno normalmente."
+  "title.addCredential": {
+    "message": "Aggiungi chiave API",
+    "description": ""
   },
-  "field": {
-    "provider": "Fornitore",
-    "label": "Note",
-    "labelPlaceholder": "es. \"Il mio numero di lavoro\" (opzionale)",
-    "apiKey": "Chiave API",
-    "apiKeyPlaceholder": "sk-...",
-    "apiKeyEditPlaceholder": "Lascia vuoto per mantenere la chiave attuale",
-    "baseUrl": "URL base personalizzato",
-    "baseUrlPlaceholder": "Lascia vuoto per utilizzare l'indirizzo predefinito del fornitore",
-    "isActive": "Attivo",
-    "lastUsedAt": "Ultimo utilizzo",
-    "createdAt": "Data di aggiunta"
+  "title.editCredential": {
+    "message": "Modifica chiave API",
+    "description": ""
   },
-  "value": {
-    "active": "Attivo",
-    "inactive": "Non attivo",
-    "never": "Mai utilizzato",
-    "default": "Predefinito"
+  "title.deleteCredential": {
+    "message": "Elimina chiave API",
+    "description": ""
   },
-  "button": {
-    "add": "Aggiungi chiave API",
-    "edit": "Modifica",
-    "test": "Test",
-    "delete": "Elimina",
-    "save": "Salva",
-    "cancel": "Annulla"
+  "description.list": {
+    "message": "Utilizza la tua chiave API del fornitore di servizi (OpenAI, Anthropic, Google, xAI, DeepSeek, Moonshot, Zhìpǔ) al posto del pool di account della piattaforma. Una volta configurato, le richieste di dialogo per il modello corrispondente verranno effettuate direttamente con la tua chiave, senza consumare crediti della piattaforma; le chiamate agli strumenti, le citazioni, i lavori e altre funzionalità come la cronologia cloud funzioneranno normalmente.",
+    "description": ""
   },
-  "message": {
-    "deleteConfirm": "Confermi di voler eliminare questa chiave API? Le conversazioni del modello {provider} torneranno a utilizzare il fornitore predefinito della piattaforma.",
-    "deleteSuccess": "Chiave API eliminata.",
-    "saveSuccess": "Chiave API salvata.",
-    "testSuccess": "Connessione riuscita.",
-    "testFailed": "Test fallito: {message}",
-    "loadFailed": "Caricamento delle credenziali fallito.",
-    "saveFailed": "Salvataggio delle credenziali fallito.",
-    "empty": "Non hai ancora configurato una chiave API personalizzata. Clicca su \"Aggiungi chiave API\" per utilizzare il tuo account del fornitore.",
-    "advancedToggle": "Mostra opzioni avanzate",
-    "baseUrlHelp": "Opzionale. Indirizzo upstream personalizzato (per OpenRouter, reverse proxy, gateway aziendale, ecc.). Solo HTTPS consentito, non può puntare a indirizzi interni/loopback."
+  "field.provider": {
+    "message": "Fornitore",
+    "description": ""
   },
-  "badge": {
-    "active": "Utilizzando la tua chiave {provider}",
-    "tooltip": "Questa conversazione utilizza la tua chiave API {provider}, i costi saranno addebitati sul tuo account {provider}, senza ridurre il saldo della piattaforma."
+  "field.label": {
+    "message": "Note",
+    "description": ""
+  },
+  "field.labelPlaceholder": {
+    "message": "es. \"Il mio numero di lavoro\" (opzionale)",
+    "description": ""
+  },
+  "field.apiKey": {
+    "message": "Chiave API",
+    "description": ""
+  },
+  "field.apiKeyPlaceholder": {
+    "message": "sk-...",
+    "description": ""
+  },
+  "field.apiKeyEditPlaceholder": {
+    "message": "Lascia vuoto per mantenere la chiave attuale",
+    "description": ""
+  },
+  "field.baseUrl": {
+    "message": "URL base personalizzato",
+    "description": ""
+  },
+  "field.baseUrlPlaceholder": {
+    "message": "Lascia vuoto per utilizzare l'indirizzo predefinito del fornitore",
+    "description": ""
+  },
+  "field.isActive": {
+    "message": "Attivo",
+    "description": ""
+  },
+  "field.lastUsedAt": {
+    "message": "Ultimo utilizzo",
+    "description": ""
+  },
+  "field.createdAt": {
+    "message": "Data di aggiunta",
+    "description": ""
+  },
+  "value.active": {
+    "message": "Attivo",
+    "description": ""
+  },
+  "value.inactive": {
+    "message": "Non attivo",
+    "description": ""
+  },
+  "value.never": {
+    "message": "Mai utilizzato",
+    "description": ""
+  },
+  "value.default": {
+    "message": "Predefinito",
+    "description": ""
+  },
+  "button.add": {
+    "message": "Aggiungi chiave API",
+    "description": ""
+  },
+  "button.edit": {
+    "message": "Modifica",
+    "description": ""
+  },
+  "button.test": {
+    "message": "Test",
+    "description": ""
+  },
+  "button.delete": {
+    "message": "Elimina",
+    "description": ""
+  },
+  "button.save": {
+    "message": "Salva",
+    "description": ""
+  },
+  "button.cancel": {
+    "message": "Annulla",
+    "description": ""
+  },
+  "message.deleteConfirm": {
+    "message": "Confermi di voler eliminare questa chiave API? Le conversazioni del modello {provider} torneranno a utilizzare il fornitore predefinito della piattaforma.",
+    "description": ""
+  },
+  "message.deleteSuccess": {
+    "message": "Chiave API eliminata.",
+    "description": ""
+  },
+  "message.saveSuccess": {
+    "message": "Chiave API salvata.",
+    "description": ""
+  },
+  "message.testSuccess": {
+    "message": "Connessione riuscita.",
+    "description": ""
+  },
+  "message.testFailed": {
+    "message": "Test fallito: {message}",
+    "description": ""
+  },
+  "message.loadFailed": {
+    "message": "Caricamento delle credenziali fallito.",
+    "description": ""
+  },
+  "message.saveFailed": {
+    "message": "Salvataggio delle credenziali fallito.",
+    "description": ""
+  },
+  "message.empty": {
+    "message": "Non hai ancora configurato una chiave API personalizzata. Clicca su \"Aggiungi chiave API\" per utilizzare il tuo account del fornitore.",
+    "description": ""
+  },
+  "message.advancedToggle": {
+    "message": "Mostra opzioni avanzate",
+    "description": ""
+  },
+  "message.baseUrlHelp": {
+    "message": "Opzionale. Indirizzo upstream personalizzato (per OpenRouter, reverse proxy, gateway aziendale, ecc.). Solo HTTPS consentito, non può puntare a indirizzi interni/loopback.",
+    "description": ""
+  },
+  "badge.active": {
+    "message": "Utilizzando la tua chiave {provider}",
+    "description": ""
+  },
+  "badge.tooltip": {
+    "message": "Questa conversazione utilizza la tua chiave API {provider}, i costi saranno addebitati sul tuo account {provider}, senza ridurre il saldo della piattaforma.",
+    "description": ""
   }
 }
