+++ src/i18n/zh-CN/common.json
@@ -703,6 +703,10 @@
     "message": "功能设置",
     "description": "设置页面中的功能设置部分的标题"
   },
+  "settings.apiKey": {
+    "message": "API Key",
+    "description": "设置页面侧边栏的 API Key 标签——用户自带（BYOK）的上游服务商 API Key 管理"
+  },
   "nav.gemini": {
     "message": "Gemini",
     "description": "Text in the website navigation bar to display the Gemini page, must remain as 'Gemini'"
